Module Tezos_raw_protocol_013_PtJakart

module Alpha_context : sig ... end

module Alpha_services : sig ... end

module Amendment : sig ... end

Amendments and proposals.

module Apply : sig ... end

This module supports advancing the ledger state by applying operations.

module Apply_results : sig ... end

module Baking : sig ... end

module Bitset : sig ... end

module Blinded_public_key_hash : sig ... end

module Block_header_repr : sig ... end

module Block_payload_hash : sig ... end

module Block_payload_repr : sig ... end

Value on which validators try to reach a consensus.

module Bond_id_repr : sig ... end

module Bootstrap_storage : sig ... end

module Cache_memory_helpers : sig ... end

module Cache_repr : sig ... end

Frequently used data should be kept in memory and persisted along a chain of blocks. The caching mechanism allows the economic protocol to declare such data and to rely on a Least Recently Used strategy to keep the cache size under a fixed limit.

module Carbonated_map : sig ... end

module Carbonated_map_costs : sig ... end

The Carbonated_map_costs module contains gas cost functions for Carbonated_map.

module Commitment_repr : sig ... end

module Commitment_storage : sig ... end

module Constants_repr : sig ... end

module Constants_services : sig ... end

module Constants_storage : sig ... end

module Contract_delegate_storage : sig ... end

module Contract_hash : sig ... end

module Contract_manager_storage : sig ... end

module Contract_repr : sig ... end

module Contract_services : sig ... end

module Contract_storage : sig ... end

module Cycle_repr : sig ... end

module Delegate_activation_storage : sig ... end

module Delegate_services : sig ... end

module Delegate_storage : sig ... end

module Dependent_bool : sig ... end

Dependent booleans

module Destination_repr : sig ... end

module Entrypoint_repr : sig ... end

module Fees_storage : sig ... end

module Fitness_repr : sig ... end

module Fixed_point_repr : sig ... end

module Frozen_deposits_storage : sig ... end

Simple abstraction from low-level storage to handle frozen deposits

module Gas_comparable_input_size : sig ... end

module Gas_input_size : sig ... end

module Gas_limit_repr : sig ... end

module Gas_monad : sig ... end

module Global_constants_costs : sig ... end

Costs function for the global table of constants.

module Global_constants_storage : sig ... end

This module represents access to a global table of constant Micheline values. Users may register a Micheline value in the table, paying the cost of storage. Once stored, contracts source code may reference this value by its hash.

module Indexable : sig ... end

module Init_storage : sig ... end

Functions to setup storage. Used by Alpha_context.prepare.

module Lazy_storage_diff : sig ... end

module Lazy_storage_kind : sig ... end

module Legacy_script_patches_for_J : sig ... end

module Level_repr : sig ... end

module Level_storage : sig ... end

module Liquidity_baking_cpmm : sig ... end

module Liquidity_baking_lqt : sig ... end

module Liquidity_baking_migration : sig ... end

module Liquidity_baking_repr : sig ... end

module Liquidity_baking_storage : sig ... end

module Local_gas_counter : sig ... end

This module exposes an API for local gas counting. It provides a set of functions for updating a gas counter without applying it on an an Alpha_context.context.

module Main : sig ... end

module Manager_repr : sig ... end

module Merkle_list : sig ... end

module Michelson_v1_gas : sig ... end

module Michelson_v1_primitives : sig ... end

module Migration_repr : sig ... end

module Misc : sig ... end

module Non_empty_string : sig ... end

module Nonce_hash : sig ... end

module Nonce_storage : sig ... end

module Operation_repr : sig ... end

module Origination_nonce : sig ... end

module Parameters_repr : sig ... end

module Path_encoding : sig ... end

module Period_repr : sig ... end

module Raw_context : sig ... end

module Raw_context_intf : sig ... end

module Raw_level_repr : sig ... end

module Receipt_repr : sig ... end

module Round_repr : sig ... end

module Sampler : sig ... end

module Sapling_repr : sig ... end

module Sapling_services : sig ... end

module Sapling_storage : sig ... end

module Sapling_validator : sig ... end

module Saturation_repr : sig ... end

module Sc_rollup_PVM_sem : sig ... end

module Sc_rollup_arith : sig ... end

module Sc_rollup_game : sig ... end

module Sc_rollup_inbox_repr : sig ... end

module Sc_rollup_operations : sig ... end

module Sc_rollup_repr : sig ... end

module Sc_rollup_storage : sig ... end

Defines storage for Smart Contract Optimistic Rollups.

module Sc_rollup_tick_repr : sig ... end

module Sc_rollups : sig ... end

module Script_cache : sig ... end

This module manages the cache for smart contracts.

module Script_comparable : sig ... end

module Script_expr_hash : sig ... end

module Script_int_repr : sig ... end

module Script_interpreter : sig ... end

This is the Michelson interpreter.

module Script_interpreter_defs : sig ... end

module Script_ir_annot : sig ... end

module Script_ir_translator : sig ... end

module Script_list : sig ... end

module Script_map : sig ... end

module Script_repr : sig ... end

module Script_set : sig ... end

module Script_string_repr : sig ... end

Strings of printable characters

module Script_tc_context : sig ... end

This module defines the typechecking context used during the translation from Michelson untyped nodes to typed nodes (Script_ir_translator). The context keeps track of the origin of the code (top-level from a contract, in a view, etc.), plus some information to allow or forbid instructions given the context (no `SELF` in a lambda for example).

module Script_tc_errors : sig ... end

module Script_tc_errors_registration : sig ... end

module Script_timestamp_repr : sig ... end

module Script_typed_ir : sig ... end

module Script_typed_ir_size : sig ... end

This module provides overapproximation of memory footprint for Michelson-related values.

module Script_typed_ir_size_costs : sig ... end

module Seed_repr : sig ... end

module Seed_storage : sig ... end

module Services_registration : sig ... end

module Skip_list_repr : sig ... end

module Slot_repr : sig ... end

module Stake_storage : sig ... end

This library provides basic operations (accessors and setters) on staking tokens.

module State_hash : sig ... end

module Storage : sig ... end

module Storage_costs : sig ... end

module Storage_description : sig ... end

module Storage_functors : sig ... end

module Storage_sigs : sig ... end

module Tez_repr : sig ... end

module Ticket_accounting : sig ... end

module Ticket_balance_key : sig ... end

This module exposes a function for generating a ticket-balance key-hash given an owner and a ticket-token. The key-hash is used for populating the global ticket-balance table that tracks ownership of tickets for different tokens.

module Ticket_balance_migration_for_j : sig ... end

module Ticket_costs : sig ... end

This module contains constants and utility functions for gas metering functions used for extracting and handling tickets for the global ticket balance table.

module Ticket_hash_builder : sig ... end

module Ticket_hash_repr : sig ... end

module Ticket_lazy_storage_diff : sig ... end

module Ticket_operations_diff : sig ... end

A module that provides functionality for extracting ticket-token differences from a list of operations.

module Ticket_scanner : sig ... end

This module provides an API for extracting tickets of arbitrary types from an OCaml values, given a type-witness.

module Ticket_storage : sig ... end

module Ticket_token : sig ... end

A module for handling ticket-tokens. A ticket-token represents the combination of a ticketer (creator of a ticket) and the content. That is, a ticket comprises a ticket-token and an amount.

module Ticket_token_map : sig ... end

A module exposing a carbonated map where keys are Ticket_token.ex_token values.

module Time_repr : sig ... end

module Token : sig ... end

The aim of this module is to manage operations involving tokens such as minting, transferring, and burning. Every constructor of the types source, container, or sink represents a kind of account that holds a given (or possibly infinite) amount of tokens.

module Tx_rollup_commitment_repr : sig ... end

module Tx_rollup_commitment_storage : sig ... end

This module introduces various functions to manipulate the storage related to commitments for transaction rollups.

module Tx_rollup_errors_repr : sig ... end

module Tx_rollup_gas : sig ... end

module Tx_rollup_hash_builder : sig ... end

module Tx_rollup_inbox_repr : sig ... end

module Tx_rollup_inbox_storage : sig ... end

Functions to manipulate transaction rollup’s inboxes.

module Tx_rollup_l2_address : sig ... end

module Tx_rollup_l2_apply : sig ... end

module Tx_rollup_l2_batch : sig ... end

module Tx_rollup_l2_context : sig ... end

module Tx_rollup_l2_context_hash : sig ... end

module Tx_rollup_l2_context_sig : sig ... end

module Tx_rollup_l2_proof : sig ... end

module Tx_rollup_l2_qty : sig ... end

module Tx_rollup_l2_storage_sig : sig ... end

module Tx_rollup_l2_verifier : sig ... end

module Tx_rollup_level_repr : sig ... end

module Tx_rollup_message_hash_repr : sig ... end

module Tx_rollup_message_repr : sig ... end

module Tx_rollup_message_result_hash_repr : sig ... end

module Tx_rollup_message_result_repr : sig ... end

module Tx_rollup_parameters : sig ... end

A module for representing and extracting typed transactional rollup parameters.

module Tx_rollup_prefixes : sig ... end

module Tx_rollup_repr : sig ... end

module Tx_rollup_reveal_repr : sig ... end

module Tx_rollup_reveal_storage : sig ... end

module Tx_rollup_services : sig ... end

module Tx_rollup_state_repr : sig ... end

module Tx_rollup_state_storage : sig ... end

A collection of functions to manipulate the state of a transaction rollup.

module Tx_rollup_storage : sig ... end

module Tx_rollup_ticket : sig ... end

This module provides various helpers to manipulate tickets, that are used by the Transaction Rollups.

module Tx_rollup_withdraw_list_hash_repr : sig ... end

module Tx_rollup_withdraw_repr : sig ... end

module Vote_repr : sig ... end

module Vote_storage : sig ... end

module Voting_period_repr : sig ... end

module Voting_period_storage : sig ... end

module Voting_services : sig ... end