Module Tezos_raw_protocol_013_PtJakart.Alpha_context

An Alpha_context.t is an immutable snapshot of the ledger state at some block height, preserving type-safety and invariants of the ledger state.

Implementation

Alpha_context.t is a wrapper over Raw_context.t, which in turn is a wrapper around Context.t from the Protocol Environment.

Lifetime of an Alpha_context

Creation, using prepare or prepare_first_block

Modification, using the operations defined in this signature

Finalization, using finalize

module type BASIC_DATA = sig ... end

type t

type context = t

type public_key = Tezos_protocol_environment_013_PtJakart.Signature.Public_key.t

type public_key_hash = Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t

type signature = Tezos_protocol_environment_013_PtJakart.Signature.t

module Slot : sig ... end

module Tez : sig ... end

module Period : sig ... end

module Timestamp : sig ... end

module Raw_level : sig ... end

module Cycle : sig ... end

module Round : sig ... end

module Gas : sig ... end

This module implements the gas subsystem of the context.

module Entrypoint : module type of Entrypoint_repr

module Script_string : module type of Script_string_repr

Strings of printable characters

module Script_int : module type of Script_int_repr

module Script_timestamp : sig ... end

module Script : sig ... end

module Constants : sig ... end

module Global_constants_storage : sig ... end

module Internal_for_tests : sig ... end

module Cache : sig ... end

module Level : sig ... end

module Fitness : sig ... end

module Nonce : sig ... end

module Seed : sig ... end

module Big_map : sig ... end

module Sapling : sig ... end

module Lazy_storage : sig ... end

module Origination_nonce : sig ... end

module Ticket_hash : sig ... end

This module re-exports functions from Ticket_hash_repr. See documentation of the functions there.

module Contract : sig ... end

module Tx_rollup_level : sig ... end

module Tx_rollup : sig ... end

This module re-exports definitions from Tx_rollup_repr and Tx_rollup_storage.

module Tx_rollup_withdraw : sig ... end

module Tx_rollup_withdraw_list_hash : sig ... end

module Tx_rollup_message_result : sig ... end

module Tx_rollup_message_result_hash : sig ... end

module Tx_rollup_commitment_hash : sig ... end

module Tx_rollup_state : sig ... end

This module re-exports definitions from Tx_rollup_state_repr and Tx_rollup_state_storage.

module Tx_rollup_reveal : sig ... end

module Tx_rollup_message : sig ... end

This module re-exports definitions from Tx_rollup_message_repr.

module Tx_rollup_message_hash : sig ... end

module Tx_rollup_inbox : sig ... end

This module re-exports definitions from Tx_rollup_inbox_repr and Tx_rollup_inbox_storage.

module Tx_rollup_commitment : sig ... end

This simply re-exports Tx_rollup_commitment_repr

module Tx_rollup_hash : sig ... end

module Tx_rollup_errors : sig ... end

module Bond_id : sig ... end

module Destination : sig ... end

This simply re-exports Destination_repr.

module Receipt : sig ... end

module Delegate : sig ... end

module Voting_period : sig ... end

module Vote : sig ... end

module Sc_rollup : sig ... end

See Sc_rollup_storage and Sc_rollup_repr.

module Block_payload : sig ... end

module Block_header : sig ... end

module Kind : sig ... end

type 'a consensus_operation_type =

| Endorsement : Kind.endorsement consensus_operation_type

| Preendorsement : Kind.preendorsement consensus_operation_type

val pp_operation_kind : Tezos_protocol_environment_013_PtJakart.Format.formatter -> 'kind consensus_operation_type -> unit

type consensus_content = {

slot : Slot.t;

level : Raw_level.t;

round : Round.t;

block_payload_hash : Block_payload_hash.t;

}

val consensus_content_encoding : consensus_content Tezos_protocol_environment_013_PtJakart.Data_encoding.t

val pp_consensus_content : Tezos_protocol_environment_013_PtJakart.Format.formatter -> consensus_content -> unit

type transaction = {

amount : Tez.tez;

parameters : Script.lazy_expr;

entrypoint : Entrypoint.t;

destination : Destination.t;

}

type origination = {

delegate : Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t option;

script : Script.t;

credit : Tez.tez;

}

type 'kind operation = {

shell : Tezos_protocol_environment_013_PtJakart.Operation.shell_header;

protocol_data : 'kind protocol_data;

}

and 'kind protocol_data = {

contents : 'kind contents_list;

signature : Tezos_protocol_environment_013_PtJakart.Signature.t option;

}

and _ contents_list =

| Single : 'kind contents -> 'kind contents_list

| Cons : 'kind Kind.manager contents * 'rest Kind.manager contents_list -> ('kind * 'rest) Kind.manager contents_list

and _ contents =

| Preendorsement : consensus_content -> Kind.preendorsement contents

| Endorsement : consensus_content -> Kind.endorsement contents

| Seed_nonce_revelation : {

level : Raw_level.t;

nonce : Nonce.t;

} -> Kind.seed_nonce_revelation contents

| Double_preendorsement_evidence : {

op1 : Kind.preendorsement operation;

op2 : Kind.preendorsement operation;

} -> Kind.double_preendorsement_evidence contents

| Double_endorsement_evidence : {

op1 : Kind.endorsement operation;

op2 : Kind.endorsement operation;

} -> Kind.double_endorsement_evidence contents

| Double_baking_evidence : {

bh1 : Block_header.t;

bh2 : Block_header.t;

} -> Kind.double_baking_evidence contents

| Activate_account : {

id : Tezos_protocol_environment_013_PtJakart.Ed25519.Public_key_hash.t;

activation_code : Blinded_public_key_hash.activation_code;

} -> Kind.activate_account contents

| Proposals : {

source : Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t;

period : int32;

proposals : Tezos_protocol_environment_013_PtJakart.Protocol_hash.t list;

} -> Kind.proposals contents

| Ballot : {

source : Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t;

period : int32;

proposal : Tezos_protocol_environment_013_PtJakart.Protocol_hash.t;

ballot : Vote.ballot;

} -> Kind.ballot contents

| Failing_noop : string -> Kind.failing_noop contents

| Manager_operation : {

source : Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t;

fee : Tez.tez;

counter : counter;

operation : 'kind manager_operation;

gas_limit : Gas.Arith.integral;

storage_limit : Tezos_protocol_environment_013_PtJakart.Z.t;

} -> 'kind Kind.manager contents

and _ manager_operation =

| Reveal : Tezos_protocol_environment_013_PtJakart.Signature.Public_key.t -> Kind.reveal manager_operation

| Transaction : transaction -> Kind.transaction manager_operation

| Origination : origination -> Kind.origination manager_operation

| Delegation : Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t option -> Kind.delegation manager_operation

| Register_global_constant : {

value : Script.lazy_expr;

} -> Kind.register_global_constant manager_operation

| Set_deposits_limit : Tez.t option -> Kind.set_deposits_limit manager_operation

| Tx_rollup_origination : Kind.tx_rollup_origination manager_operation

| Tx_rollup_submit_batch : {

tx_rollup : Tx_rollup.t;

content : string;

burn_limit : Tez.tez option;

} -> Kind.tx_rollup_submit_batch manager_operation

| Tx_rollup_commit : {

tx_rollup : Tx_rollup.t;

commitment : Tx_rollup_commitment.Full.t;

} -> Kind.tx_rollup_commit manager_operation

| Tx_rollup_return_bond : {

tx_rollup : Tx_rollup.t;

} -> Kind.tx_rollup_return_bond manager_operation

| Tx_rollup_finalize_commitment : {

tx_rollup : Tx_rollup.t;

} -> Kind.tx_rollup_finalize_commitment manager_operation

| Tx_rollup_remove_commitment : {

tx_rollup : Tx_rollup.t;

} -> Kind.tx_rollup_remove_commitment manager_operation

| Tx_rollup_rejection : {

tx_rollup : Tx_rollup.t;

level : Tx_rollup_level.t;

message : Tx_rollup_message.t;

message_position : int;

message_path : Tx_rollup_inbox.Merkle.path;

message_result_hash : Tx_rollup_message_result_hash.t;

message_result_path : Tx_rollup_commitment.Merkle.path;

previous_message_result : Tx_rollup_message_result.t;

previous_message_result_path : Tx_rollup_commitment.Merkle.path;

proof : Tx_rollup_l2_proof.t;

} -> Kind.tx_rollup_rejection manager_operation

| Tx_rollup_dispatch_tickets : {

tx_rollup : Tx_rollup.t;

level : Tx_rollup_level.t;

context_hash : Tezos_protocol_environment_013_PtJakart.Context_hash.t;

message_index : int;

message_result_path : Tx_rollup_commitment.Merkle.path;

tickets_info : Tx_rollup_reveal.t list;

} -> Kind.tx_rollup_dispatch_tickets manager_operation

| Transfer_ticket : {

contents : Script.lazy_expr;

ty : Script.lazy_expr;

ticketer : Contract.t;

amount : Tezos_protocol_environment_013_PtJakart.Z.t;

destination : Contract.t;

entrypoint : Entrypoint.t;

} -> Kind.transfer_ticket manager_operation

| Sc_rollup_originate : {

kind : Sc_rollup.Kind.t;

boot_sector : string;

} -> Kind.sc_rollup_originate manager_operation

| Sc_rollup_add_messages : {

rollup : Sc_rollup.t;

messages : string list;

} -> Kind.sc_rollup_add_messages manager_operation

| Sc_rollup_cement : {

rollup : Sc_rollup.t;

commitment : Sc_rollup.Commitment_hash.t;

} -> Kind.sc_rollup_cement manager_operation

| Sc_rollup_publish : {

rollup : Sc_rollup.t;

commitment : Sc_rollup.Commitment.t;

} -> Kind.sc_rollup_publish manager_operation

and counter = Tezos_protocol_environment_013_PtJakart.Z.t

type packed_manager_operation =

| Manager : 'kind manager_operation -> packed_manager_operation

type packed_contents =

| Contents : 'kind contents -> packed_contents

type packed_contents_list =

| Contents_list : 'kind contents_list -> packed_contents_list

type packed_protocol_data =

| Operation_data : 'kind protocol_data -> packed_protocol_data

type packed_operation = {

shell : Tezos_protocol_environment_013_PtJakart.Operation.shell_header;

protocol_data : packed_protocol_data;

}

val manager_kind : 'kind manager_operation -> 'kind Kind.manager

module Operation : sig ... end

module Stake_distribution : sig ... end

module Commitment : sig ... end

module Bootstrap : sig ... end

module Migration : sig ... end

val prepare_first_block : Tezos_protocol_environment_013_PtJakart.Context.t -> typecheck: (context -> Script.t -> ((Script.t * Lazy_storage.diffs option) * context) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t) -> level:Tezos_protocol_environment_013_PtJakart.Int32.t -> timestamp:Tezos_protocol_environment_013_PtJakart.Time.t -> context Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Create an Alpha_context.t from an untyped context (first block in the chain only).

val prepare : Tezos_protocol_environment_013_PtJakart.Context.t -> level:Tezos_protocol_environment_013_PtJakart.Int32.t -> predecessor_timestamp:Tezos_protocol_environment_013_PtJakart.Time.t -> timestamp:Tezos_protocol_environment_013_PtJakart.Time.t -> (context * Receipt.balance_updates * Migration.origination_result list) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Create an Alpha_context.t from an untyped context.

val activate : context -> Tezos_protocol_environment_013_PtJakart.Protocol_hash.t -> context Tezos_protocol_environment_013_PtJakart.Lwt.t

val reset_internal_nonce : context -> context

val fresh_internal_nonce : context -> (context * int) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult

val record_internal_nonce : context -> int -> context

val internal_nonce_already_recorded : context -> int -> bool

val description : context Storage_description.t

val finalize : ?commit_message:string -> context -> Fitness.raw -> Tezos_protocol_environment_013_PtJakart.Updater.validation_result

Finalize an Alpha_context.t, producing a validation_result.

val current_context : context -> Tezos_protocol_environment_013_PtJakart.Context.t

Should only be used by Main.current_context to return a context usable for RPCs

val record_non_consensus_operation_hash : context -> Tezos_protocol_environment_013_PtJakart.Operation_hash.t -> context

val non_consensus_operations : context -> Tezos_protocol_environment_013_PtJakart.Operation_hash.t list

module Parameters : sig ... end

module Liquidity_baking : sig ... end

module Ticket_balance : sig ... end

This module re-exports functions from Ticket_storage. See documentation of the functions there.

module First_level_of_protocol : sig ... end

module Consensus : sig ... end

module Token : sig ... end

See 'token.mli' for more explanation.

module Fees : sig ... end