Module Tezos_raw_protocol_013_PtJakart.Baking

type Tezos_protocol_environment_013_PtJakart.Error_monad.error +=

| Insufficient_endorsing_power of {

endorsing_power : int;

consensus_threshold : int;

}

type ordered_slots = private Alpha_context.Slot.t list

val endorsing_rights : Alpha_context.context -> Alpha_context.Level.t -> (Alpha_context.context * ordered_slots Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.Map.t) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

For a given level computes who has the right to include an endorsement in the next block.

returns
map from delegates with such rights to their endorsing slots, in increasing order.

This function is only used by the 'validators' RPC.

val endorsing_rights_by_first_slot : Alpha_context.context -> Alpha_context.Level.t -> (Alpha_context.context * (Alpha_context.public_key * Alpha_context.public_key_hash * int) Alpha_context.Slot.Map.t) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Computes endorsing rights for a given level.

returns
map from allocated first slots to their owner's public key, public key hash, and endorsing power.

val bonus_baking_reward : Alpha_context.context -> endorsing_power:int -> Alpha_context.Tez.t Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult

Computes the bonus baking reward depending on the endorsing power.

val baking_rights : Alpha_context.context -> Alpha_context.Level.t -> Alpha_context.public_key Misc.lazy_list

baking_rights ctxt level is the lazy list of contract's public key hashes that are allowed to propose for level at each round.