Module Tezos_raw_protocol_013_PtJakart.Carbonated_map

module type S = sig ... end

An in-memory data-structure for a key-value map where all operations account for gas costs.

module type COMPARABLE = sig ... end

A module type for comparable values that also includes a cost function for metering gas depending on the cost of comparing values.

module Make (O : COMPARABLE) : S with type key := O.t

A functor for building gas metered maps.