Parameter Make.O

include Tezos_protocol_environment_013_PtJakart.Compare.COMPARABLE

type t

val compare : t -> t -> int

val compare_cost : t -> Alpha_context.Gas.cost

compare_cost k returns the cost of comparing the given key k with another value of the same type.