Module Tezos_raw_protocol_013_PtJakart.Delegate_activation_storage

val is_inactive : Raw_context.t -> Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t -> bool Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

val last_cycle_before_deactivation : Raw_context.t -> Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t -> Cycle_repr.t Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

last_cycle_before_deactivation ctxt delegate is the cycle at which the delegate is scheduled to become inactive.

val set_inactive : Raw_context.t -> Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t -> Raw_context.t Tezos_protocol_environment_013_PtJakart.Lwt.t

set_inactive context delegate adds delegate to the set of inactive * contracts.

val set_active : Raw_context.t -> Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t -> (Raw_context.t * bool) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

set_active ctxt delegate returns a pair (new_ctxt, is_inactive) where:

new_ctxt is a new context, updated from ctxt, where the delegate's last active cycle has been updated

is_inactive represents the state of delegate, prior to the update.