Module Tezos_raw_protocol_013_PtJakart.Gas_comparable_input_size

Gas_input_size includes the definitions for the different sizes used in the gas models of the protocol. They do not always represent memory sizes, but rather they can be seen as an information size. They are tailored to the models that use them, and should not be used for anything other than gas computation.

Gas_comparable_input_size is the restriction of Gas_input_size to comparable types.

type t = int

type micheline_size = {

traversal : t;

int_bytes : t;

string_bytes : t;

}

val encoding : t Tezos_protocol_environment_013_PtJakart.Data_encoding.encoding

val micheline_size_encoding : micheline_size Tezos_protocol_environment_013_PtJakart.Data_encoding.encoding

val zero : t

val add : t -> t -> t

val pp : Tezos_protocol_environment_013_PtJakart.Format.formatter -> t -> unit

val pp_micheline_size : Tezos_protocol_environment_013_PtJakart.Format.formatter -> micheline_size -> unit

val to_int : t -> int

val of_int : int -> t

val integer : 'a Alpha_context.Script_int.num -> t

val string : string -> t

val script_string : Alpha_context.Script_string.t -> t

val bytes : Tezos_protocol_environment_013_PtJakart.Bytes.t -> t

val mutez : Alpha_context.Tez.tez -> t

val timestamp : Alpha_context.Script_timestamp.t -> t

val size_of_comparable_value : 'a Script_typed_ir.comparable_ty -> 'a -> t