Module Tezos_raw_protocol_013_PtJakart.Parameters_repr

This module defines protocol parameters, i.e. constants regulating the behaviour of the blockchain under the protocol.

type bootstrap_account = {

public_key_hash : Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t;

public_key : Tezos_protocol_environment_013_PtJakart.Signature.Public_key.t option;

amount : Tez_repr.t;

}

An implict contract (account) initially existing on a chain since genesis.

type bootstrap_contract = {

delegate : Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t option;

amount : Tez_repr.t;

script : Script_repr.t;

}

An originated contract initially existing on a chain since genesis.

type t = {

bootstrap_accounts : bootstrap_account list;

bootstrap_contracts : bootstrap_contract list;

commitments : Commitment_repr.t list;

constants : Constants_repr.parametric;

security_deposit_ramp_up_cycles : int option;

no_reward_cycles : int option;

}

Protocol parameters define some constants regulating behaviour of the chain.

val encoding : t Tezos_protocol_environment_013_PtJakart.Data_encoding.t

val check_params : t -> unit Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult