Module type Path_encoding.S

type t

val to_path : t -> string list -> string list

to_path t postfix returns the context path name for t postfixed with postfix

val of_path : string list -> t option

of_path path parses path as a context path name for t

val path_length : int

Directory levels of the path encoding of t