Module Sc_rollup_repr.Number_of_messages

Number of messages consumed by a single commitment. This represents a claim about the shape of the Inbox, which can be disputed as part of a commitment dispute.

See also Sc_rollup_repr.Commitments.

type t

val min_int : int32

val max_int : int32

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val encoding : t Tezos_protocol_environment_013_PtJakart.Data_encoding.t

val to_int32 : t -> int32

val of_int32 : int32 -> t option