Module type PVM.S

val name : string

val parse_boot_sector : string -> boot_sector option

val pp_boot_sector : Tezos_protocol_environment_013_PtJakart.Format.formatter -> boot_sector -> unit

include Sc_rollup_PVM_sem.S

type state

The state of the PVM denotes a state of the rollup.

type proof

During interactive rejection games, a player may need to provide a proof that a given execution step is valid.

val proof_encoding : proof Tezos_protocol_environment_013_PtJakart.Data_encoding.t

type context

A state is initialized in a given context.

type hash = Alpha_context.Sc_rollup.State_hash.t

A commitment hash characterized the contents of the state.

val proof_start_state : proof -> hash

proof_start_state proof returns the initial state hash of the proof execution step.

val proof_stop_state : proof -> hash

proof_stop_state proof returns the final state hash of the proof execution step.

val state_hash : state -> hash Tezos_protocol_environment_013_PtJakart.Lwt.t

state_hash state returns a compressed representation of state.

val initial_state : context -> string -> state Tezos_protocol_environment_013_PtJakart.Lwt.t

initial_state context is the state of the PVM before booting. It must be such that state_hash state = Commitment_hash.zero. Any context should be enough to create an initial state.

val is_input_state : state -> (Alpha_context.Raw_level.t * Tezos_protocol_environment_013_PtJakart.Z.t) option Tezos_protocol_environment_013_PtJakart.Lwt.t

is_input_state state returns Some (level, counter) if state is waiting for the input message that comes next to the message numbered counter in the inbox of a given level.

val set_input : Sc_rollup_PVM_sem.input -> state -> state Tezos_protocol_environment_013_PtJakart.Lwt.t

set_input level n msg state sets msg in state as the next message to be processed. This input message is assumed to be the number n in the inbox messages at the given level. The input message must be the message next to the previous message processed by the rollup.

val eval : state -> state Tezos_protocol_environment_013_PtJakart.Lwt.t

eval s0 returns a state s1 resulting from the execution of an atomic step of the rollup at state s0.

val verify_proof : input:Sc_rollup_PVM_sem.input option -> proof -> bool Tezos_protocol_environment_013_PtJakart.Lwt.t

verify_proof input proof returns true iff the proof is valid. If the state is an input state, input is the hash of the input message externally provided to the evaluation function.