Module Script_interpreter.Internals

Internally, the interpretation loop uses a local gas counter.

val next : Script_typed_ir.logger option -> (Local_gas_counter.outdated_context * step_constants) -> Local_gas_counter.local_gas_counter -> ('a, 's, 'r, 'f) Script_typed_ir.continuation -> 'a -> 's -> ('r * 'f * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

next logger (ctxt, step_constants) local_gas_counter ks accu stack is an internal function which interprets the continuation ks to execute the interpreter on the current A-stack.

val step : (Local_gas_counter.outdated_context * step_constants) -> Local_gas_counter.local_gas_counter -> ('a, 's, 'r, 'f) Script_typed_ir.kinstr -> 'a -> 's -> ('r * 'f * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t