Module Tezos_raw_protocol_013_PtJakart.Script_interpreter_defs

module Interp_costs = Michelson_v1_gas.Cost_of.Interpreter

val cost_of_instr : 'a 's 'r 'f. ('a, 's, 'r, 'f) Script_typed_ir.kinstr -> 'a -> 's -> Alpha_context.Gas.cost

val cost_of_control : 'a 's 'r 'f. ('a, 's, 'r, 'f) Script_typed_ir.continuation -> Alpha_context.Gas.cost

val consume_instr : Local_gas_counter.local_gas_counter -> ('a, 'b, 'c, 'd) Script_typed_ir.kinstr -> 'a -> 'b -> Local_gas_counter.local_gas_counter option

val consume_control : Local_gas_counter.local_gas_counter -> ('a, 'b, 'c, 'd) Script_typed_ir.continuation -> Local_gas_counter.local_gas_counter option

val log_entry : Script_typed_ir.logger -> Local_gas_counter.outdated_context -> Local_gas_counter.local_gas_counter -> ('a, 'b, 'c, 'd) Script_typed_ir.kinstr -> 'a -> 'b -> unit

val log_exit : Script_typed_ir.logger -> Local_gas_counter.outdated_context -> Local_gas_counter.local_gas_counter -> ('a, 'b) Script_typed_ir.kinfo -> ('c, 'd, 'e, 'f) Script_typed_ir.kinstr -> 'c -> 'd -> unit

val log_control : Script_typed_ir.logger -> ('a, 'b, 'c, 'd) Script_typed_ir.continuation -> unit

val get_log : Script_typed_ir.logger option -> (Script_typed_ir.execution_trace option, Tezos_protocol_environment_013_PtJakart.Error_monad.error Tezos_protocol_environment_013_PtJakart.Error_monad.trace) Tezos_protocol_environment_013_PtJakart.Pervasives.result Tezos_protocol_environment_013_PtJakart.Lwt.t

val log_kinstr : Script_typed_ir.logger -> ('a, 'b, 'c, 'd) Script_typed_ir.kinstr -> ('a, 'b, 'c, 'd) Script_typed_ir.kinstr

val log_next_kinstr : Script_typed_ir.logger -> ('a, 'b, 'c, 'd) Script_typed_ir.kinstr -> ('a, 'b, 'c, 'd) Script_typed_ir.kinstr

val id : 'a -> 'a

val kundip : 'a 's 'e 'z 'c 'u 'd 'w 'b 't. ('a, 's, 'e, 'z, 'c, 'u, 'd, 'w) Script_typed_ir.stack_prefix_preservation_witness -> 'c -> 'u -> ('d, 'w, 'b, 't) Script_typed_ir.kinstr -> 'a * 's * ('e, 'z, 'b, 't) Script_typed_ir.kinstr

val apply : Local_gas_counter.outdated_context -> Local_gas_counter.local_gas_counter -> ('a, 'b) Script_typed_ir.ty -> 'a -> (('a, 'c) Script_typed_ir.pair, 'd) Script_typed_ir.lambda -> (('c, 'd) Script_typed_ir.lambda * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter, Tezos_protocol_environment_013_PtJakart.Error_monad.error Tezos_protocol_environment_013_PtJakart.Error_monad.trace) Tezos_protocol_environment_013_PtJakart.Pervasives.result Tezos_protocol_environment_013_PtJakart.Lwt.t

val transfer : (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> Alpha_context.Tez.tez -> Alpha_context.Script.location -> ('a, 'b) Script_typed_ir.ty -> 'a -> Alpha_context.Destination.t -> Alpha_context.Entrypoint.t -> (Script_typed_ir.operation * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter, Tezos_protocol_environment_013_PtJakart.Error_monad.error Tezos_protocol_environment_013_PtJakart.Error_monad.trace) Tezos_protocol_environment_013_PtJakart.Pervasives.result Tezos_protocol_environment_013_PtJakart.Lwt.t

val create_contract : (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> ('a, 'b) Script_typed_ir.ty -> Alpha_context.Script.expr -> Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t option -> Alpha_context.Tez.tez -> 'a -> (Script_typed_ir.operation * Alpha_context.Contract.t * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter, Tezos_protocol_environment_013_PtJakart.Error_monad.error Tezos_protocol_environment_013_PtJakart.Error_monad.trace) Tezos_protocol_environment_013_PtJakart.Pervasives.result Tezos_protocol_environment_013_PtJakart.Lwt.t

create_contract (ctxt, sc) gas storage_ty code delegate credit init creates an origination operation for a contract represented by code, some initial credit (withdrawn from the contract being executed), and an initial storage init of type storage_ty.

val unpack : Tezos_raw_protocol_013_PtJakart__Alpha_context.t -> ty:('a, 'b) Script_typed_ir.ty -> bytes:bytes -> ('a option * Tezos_raw_protocol_013_PtJakart__Alpha_context.t, Tezos_protocol_environment_013_PtJakart.Error_monad.error Tezos_protocol_environment_013_PtJakart.Error_monad.trace) Tezos_protocol_environment_013_PtJakart.Pervasives.result Tezos_protocol_environment_013_PtJakart.Lwt.t

val interp_stack_prefix_preserving_operation : 'a 's 'b 't 'c 'u 'd 'w 'result. ('a -> 's -> ('b * 't) * 'result) -> ('a, 's, 'b, 't, 'c, 'u, 'd, 'w) Script_typed_ir.stack_prefix_preservation_witness -> 'c -> 'u -> ('d * 'w) * 'result

type ('a, 's, 'b, 't, 'r, 'f) step_type = (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> ('a, 's, 'b, 't) Script_typed_ir.kinstr -> ('b, 't, 'r, 'f) Script_typed_ir.continuation -> 'a -> 's -> ('r * 'f * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'f, 'g, 'h, 'm, 'n, 'o) kmap_exit_type = (('c, 'd, 'e, 'f) Script_typed_ir.continuation -> ('a, 'b, 'g, 'h) Script_typed_ir.continuation) -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> (('m * 'n, 'c * 'd, 'o, 'c * 'd) Script_typed_ir.kinstr * ('m * 'n) list * ('m, 'o) Script_typed_ir.map * 'm) -> (('m, 'o) Script_typed_ir.map, 'c * 'd, 'e, 'f) Script_typed_ir.continuation -> 'o -> ('a * 'b) -> ('g * 'h * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'j, 'k) kmap_enter_type = (('a, 'b * 'c, 'd, 'e) Script_typed_ir.continuation -> ('a, 'b * 'c, 'd, 'e) Script_typed_ir.continuation) -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> (('j * 'k, 'b * 'c, 'a, 'b * 'c) Script_typed_ir.kinstr * ('j * 'k) list * ('j, 'a) Script_typed_ir.map) -> (('j, 'a) Script_typed_ir.map, 'b * 'c, 'd, 'e) Script_typed_ir.continuation -> 'b -> 'c -> ('d * 'e * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'i, 'j) klist_exit_type = (('a, 'b, 'c, 'd) Script_typed_ir.continuation -> ('a, 'b, 'c, 'd) Script_typed_ir.continuation) -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> (('i, 'a * 'b, 'j, 'a * 'b) Script_typed_ir.kinstr * 'i list * 'j list * int) -> ('j Script_typed_ir.boxed_list, 'a * 'b, 'c, 'd) Script_typed_ir.continuation -> 'j -> ('a * 'b) -> ('c * 'd * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'j) klist_enter_type = (('b, 'a * 'c, 'd, 'e) Script_typed_ir.continuation -> ('b, 'a * 'c, 'd, 'e) Script_typed_ir.continuation) -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> (('j, 'a * 'c, 'b, 'a * 'c) Script_typed_ir.kinstr * 'j list * 'b list * int) -> ('b Script_typed_ir.boxed_list, 'a * 'c, 'd, 'e) Script_typed_ir.continuation -> 'a -> 'c -> ('d * 'e * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'f, 'g) kloop_in_left_type = (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> ('c, 'd, 'e, 'f) Script_typed_ir.continuation -> ('a, 'g, 'c, 'd) Script_typed_ir.kinstr -> ('b, 'g, 'e, 'f) Script_typed_ir.continuation -> ('a, 'b) Script_typed_ir.union -> 'g -> ('e * 'f * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'r, 'f, 's) kloop_in_type = (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> ('b, 'c, 'r, 'f) Script_typed_ir.continuation -> ('a, 's, 'b, 'c) Script_typed_ir.kinstr -> ('a, 's, 'r, 'f) Script_typed_ir.continuation -> bool -> ('a * 's) -> ('r * 'f * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 's, 'r, 'f) kiter_type = (('a, 's, 'r, 'f) Script_typed_ir.continuation -> ('a, 's, 'r, 'f) Script_typed_ir.continuation) -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> (('b, 'a * 's, 'a, 's) Script_typed_ir.kinstr * 'b list) -> ('a, 's, 'r, 'f) Script_typed_ir.continuation -> 'a -> 's -> ('r * 'f * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'f, 'g, 'h) ilist_map_type = (('a, 'b, 'c, 'd) Script_typed_ir.continuation -> ('a, 'b, 'c, 'd) Script_typed_ir.continuation) -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> (('e, 'a * 'b, 'f, 'a * 'b) Script_typed_ir.kinstr * ('f Script_typed_ir.boxed_list, 'a * 'b, 'g, 'h) Script_typed_ir.kinstr) -> ('g, 'h, 'c, 'd) Script_typed_ir.continuation -> 'e Script_typed_ir.boxed_list -> ('a * 'b) -> ('c * 'd * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'f, 'g) ilist_iter_type = (('a, 'b, 'c, 'd) Script_typed_ir.continuation -> ('a, 'b, 'c, 'd) Script_typed_ir.continuation) -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> (('e, 'a * 'b, 'a, 'b) Script_typed_ir.kinstr * ('a, 'b, 'f, 'g) Script_typed_ir.kinstr) -> ('f, 'g, 'c, 'd) Script_typed_ir.continuation -> 'e Script_typed_ir.boxed_list -> ('a * 'b) -> ('c * 'd * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'f, 'g) iset_iter_type = (('a, 'b, 'c, 'd) Script_typed_ir.continuation -> ('a, 'b, 'c, 'd) Script_typed_ir.continuation) -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> (('e, 'a * 'b, 'a, 'b) Script_typed_ir.kinstr * ('a, 'b, 'f, 'g) Script_typed_ir.kinstr) -> ('f, 'g, 'c, 'd) Script_typed_ir.continuation -> 'e Script_typed_ir.set -> ('a * 'b) -> ('c * 'd * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'f, 'g, 'h, 'i) imap_map_type = (('a, 'b, 'c, 'd) Script_typed_ir.continuation -> ('a, 'b, 'c, 'd) Script_typed_ir.continuation) -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> (('e * 'f, 'a * 'b, 'g, 'a * 'b) Script_typed_ir.kinstr * (('e, 'g) Script_typed_ir.map, 'a * 'b, 'h, 'i) Script_typed_ir.kinstr) -> ('h, 'i, 'c, 'd) Script_typed_ir.continuation -> ('e, 'f) Script_typed_ir.map -> ('a * 'b) -> ('c * 'd * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'f, 'g, 'h) imap_iter_type = (('a, 'b, 'c, 'd) Script_typed_ir.continuation -> ('a, 'b, 'c, 'd) Script_typed_ir.continuation) -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> (('e * 'f, 'a * 'b, 'a, 'b) Script_typed_ir.kinstr * ('a, 'b, 'g, 'h) Script_typed_ir.kinstr) -> ('g, 'h, 'c, 'd) Script_typed_ir.continuation -> ('e, 'f) Script_typed_ir.map -> ('a * 'b) -> ('c * 'd * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'f) imul_teznat_type = Script_typed_ir.logger option -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> ((Alpha_context.Tez.t, 'a) Script_typed_ir.kinfo * (Alpha_context.Tez.t, 'b, 'c, 'd) Script_typed_ir.kinstr) -> ('c, 'd, 'e, 'f) Script_typed_ir.continuation -> Alpha_context.Tez.t -> (Alpha_context.Script_int.n Alpha_context.Script_int.num * 'b) -> ('e * 'f * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter, Tezos_protocol_environment_013_PtJakart.Error_monad.error Tezos_protocol_environment_013_PtJakart.Error_monad.trace) Tezos_protocol_environment_013_PtJakart.Pervasives.result Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'f) imul_nattez_type = Script_typed_ir.logger option -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> ((Alpha_context.Script_int.n Alpha_context.Script_int.num, 'a) Script_typed_ir.kinfo * (Alpha_context.Tez.t, 'b, 'c, 'd) Script_typed_ir.kinstr) -> ('c, 'd, 'e, 'f) Script_typed_ir.continuation -> Alpha_context.Script_int.n Alpha_context.Script_int.num -> (Alpha_context.Tez.t * 'b) -> ('e * 'f * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter, Tezos_protocol_environment_013_PtJakart.Error_monad.error Tezos_protocol_environment_013_PtJakart.Error_monad.trace) Tezos_protocol_environment_013_PtJakart.Pervasives.result Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'f) ilsl_nat_type = Script_typed_ir.logger option -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> ((Alpha_context.Script_int.n Alpha_context.Script_int.num, 'a) Script_typed_ir.kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 'b, 'c, 'd) Script_typed_ir.kinstr) -> ('c, 'd, 'e, 'f) Script_typed_ir.continuation -> Alpha_context.Script_int.n Alpha_context.Script_int.num -> (Alpha_context.Script_int.n Alpha_context.Script_int.num * 'b) -> ('e * 'f * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter, Tezos_protocol_environment_013_PtJakart.Error_monad.error Tezos_protocol_environment_013_PtJakart.Error_monad.trace) Tezos_protocol_environment_013_PtJakart.Pervasives.result Tezos_protocol_environment_013_PtJakart.Lwt.t

type ('a, 'b, 'c, 'd, 'e, 'f) ilsr_nat_type = Script_typed_ir.logger option -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> ((Alpha_context.Script_int.n Alpha_context.Script_int.num, 'a) Script_typed_ir.kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 'b, 'c, 'd) Script_typed_ir.kinstr) -> ('c, 'd, 'e, 'f) Script_typed_ir.continuation -> Alpha_context.Script_int.n Alpha_context.Script_int.num -> (Alpha_context.Script_int.n Alpha_context.Script_int.num * 'b) -> ('e * 'f * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter, Tezos_protocol_environment_013_PtJakart.Error_monad.error Tezos_protocol_environment_013_PtJakart.Error_monad.trace) Tezos_protocol_environment_013_PtJakart.Pervasives.result Tezos_protocol_environment_013_PtJakart.Lwt.t

type ifailwith_type = {

ifailwith : 'a 'ac 'b. Script_typed_ir.logger option -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> Alpha_context.Script.location -> ('a, 'ac) Script_typed_ir.ty -> 'a -> ('b, Tezos_protocol_environment_013_PtJakart.Error_monad.error Tezos_protocol_environment_013_PtJakart.Error_monad.trace) Tezos_protocol_environment_013_PtJakart.Pervasives.result Tezos_protocol_environment_013_PtJakart.Lwt.t;

}

type ('a, 'b, 'c, 'd, 'e, 'f, 'g) iexec_type = Script_typed_ir.logger option -> (Local_gas_counter.outdated_context * Script_typed_ir.step_constants) -> Local_gas_counter.local_gas_counter -> ('a, 'b, 'c, 'd) Script_typed_ir.kinstr -> ('c, 'd, 'e, 'f) Script_typed_ir.continuation -> 'g -> (('g, 'a) Script_typed_ir.lambda * 'b) -> ('e * 'f * Local_gas_counter.outdated_context * Local_gas_counter.local_gas_counter) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t