Module Storage.Contract

Storage from this submodule must only be accessed through the module `Contract`.

module Global_counter : Simple_single_data_storage with type value = Tezos_protocol_environment_013_PtJakart.Z.t

val fold : Raw_context.t -> order:[ `Sorted | `Undefined ] -> init:'a -> f:(Contract_repr.t -> 'a -> 'a Tezos_protocol_environment_013_PtJakart.Lwt.t) -> 'a Tezos_protocol_environment_013_PtJakart.Lwt.t

The domain of alive contracts

val list : Raw_context.t -> Contract_repr.t list Tezos_protocol_environment_013_PtJakart.Lwt.t

module Spendable_balance : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tez_repr.t and type t := Raw_context.t

The tez possessed by a contract and that can be used. A contract may also possess tez in frozen deposits. Empty balances (of zero tez) are only allowed for originated contracts, not for implicit ones.

module Missed_endorsements : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = missed_endorsements_info and type t := Raw_context.t

If the value is not set, the delegate didn't miss any endorsing opportunity. If it is set, this value is a record of type missed_endorsements_info, where:

module Manager : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Manager_repr.t and type t := Raw_context.t

The manager of a contract

module Delegate : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t and type t := Raw_context.t

The delegate of a contract, if any.

module Delegated : Storage_sigs.Data_set_storage with type elt = Contract_repr.t and type t = Raw_context.t * Contract_repr.t

All contracts (implicit and originated) that are delegated, if any

module Frozen_deposits : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = deposits and type t := Raw_context.t

The part of a delegate balance that can't be used. The total balance is frozen_deposits.current_amount + balance. It also stores the initial frozen balance in frozen_deposits.initial_amount. We have current_amount <= initial_amount and current_amount < initial_amount iff the delegate was slashed.

module Frozen_deposits_limit : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tez_repr.t and type t := Raw_context.t

If there is a value, the frozen balance for the contract won't exceed it (starting in preserved_cycles + 1).

module Inactive_delegate : Storage_sigs.Data_set_storage with type elt = Contract_repr.t and type t = Raw_context.t

module Delegate_last_cycle_before_deactivation : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Cycle_repr.t and type t := Raw_context.t

The last cycle where the delegate is considered active; that is, at the next cycle it will be considered inactive.

module Counter : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tezos_protocol_environment_013_PtJakart.Z.t and type t := Raw_context.t

module Code : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Contract_repr.t and type value = Script_repr.lazy_expr and type t := Raw_context.t

module Storage : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Contract_repr.t and type value = Script_repr.lazy_expr and type t := Raw_context.t

module Used_storage_space : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tezos_protocol_environment_013_PtJakart.Z.t and type t := Raw_context.t

Current storage space in bytes. Includes code, global storage and big map elements.

module Paid_storage_space : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tezos_protocol_environment_013_PtJakart.Z.t and type t := Raw_context.t

Maximal space available without needing to burn new fees.

module Frozen_bonds : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Bond_id_repr.t and type value = Tez_repr.t and type t := Raw_context.t * Contract_repr.t

Associates a contract and a bond_id with a bond, i.e. an amount of tez that is frozen.

val fold_bond_ids : (Raw_context.t * Contract_repr.t) -> order:[ `Sorted | `Undefined ] -> init:'a -> f:(Bond_id_repr.t -> 'a -> 'a Tezos_protocol_environment_013_PtJakart.Lwt.t) -> 'a Tezos_protocol_environment_013_PtJakart.Lwt.t

module Total_frozen_bonds : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tez_repr.t and type t := Raw_context.t

Associates a contract with the total of all its frozen bonds.