Module Contract.Missed_endorsements

If the value is not set, the delegate didn't miss any endorsing opportunity. If it is set, this value is a record of type missed_endorsements_info, where:

remaining_slots is the difference between the maximum number of slots that can be missed and the number of missed slots; therefore, when the number is positive, it represents the number of slots that a delegate can still miss before forfeiting its endorsing rewards for the current cycle; when the number is zero it means rewards are not lost, but no further slots can be missed anymore;

missed_levels represents the number of missed levels (for endorsing).

include Storage_sigs.Non_iterable_indexed_data_storage with type key = Contract_repr.t with type value = missed_endorsements_info with type t := Raw_context.t

type context = Raw_context.t

type key = Contract_repr.t

An abstract type for keys

type value = missed_endorsements_info

The type of values

val mem : context -> key -> bool Tezos_protocol_environment_013_PtJakart.Lwt.t

Tells if a given key is already bound to a storage bucket

val get : context -> key -> value Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Retrieve a value from the storage bucket at a given key ; returns Storage_errorMissing_key if the key is not set ; returns Storage_errorCorrupted_data if the deserialisation fails.

val find : context -> key -> value option Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Retrieve a value from the storage bucket at a given key ; returns None if the value is not set ; returns Storage_error Corrupted_data if the deserialisation fails.

val update : context -> key -> value -> Raw_context.t Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Updates the content of a bucket ; returns A Storage_Error Missing_key if the value does not exists.

val init : context -> key -> value -> Raw_context.t Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Allocates a storage bucket at the given key and initializes it ; returns a Storage_errorExisting_key if the bucket exists.

val add : context -> key -> value -> Raw_context.t Tezos_protocol_environment_013_PtJakart.Lwt.t

Allocates a storage bucket at the given key and initializes it with a value ; just updates it if the bucket exists.

val add_or_remove : context -> key -> value option -> Raw_context.t Tezos_protocol_environment_013_PtJakart.Lwt.t

When the value is Some v, allocates the data and initializes it with v ; just updates it if the bucket exists. When the value is None, delete the storage bucket when the value ; does nothing if the bucket does not exists.

val remove_existing : context -> key -> Raw_context.t Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Delete a storage bucket and its contents ; returns a Storage_errorMissing_key if the bucket does not exists.

val remove : context -> key -> Raw_context.t Tezos_protocol_environment_013_PtJakart.Lwt.t

Removes a storage bucket and its contents ; does nothing if the bucket does not exists.

val clear : context -> Raw_context.t Tezos_protocol_environment_013_PtJakart.Lwt.t

Empties all the keys and associated data.

val keys : context -> key list Tezos_protocol_environment_013_PtJakart.Lwt.t

Lists all the keys.

val bindings : context -> (key * value) list Tezos_protocol_environment_013_PtJakart.Lwt.t

Lists all the keys and associated data.

val fold : context -> order:[ `Sorted | `Undefined ] -> init:'a -> f:(key -> value -> 'a -> 'a Tezos_protocol_environment_013_PtJakart.Lwt.t) -> 'a Tezos_protocol_environment_013_PtJakart.Lwt.t

Iterates over all the keys and associated data.

val fold_keys : context -> order:[ `Sorted | `Undefined ] -> init:'a -> f:(key -> 'a -> 'a Tezos_protocol_environment_013_PtJakart.Lwt.t) -> 'a Tezos_protocol_environment_013_PtJakart.Lwt.t

Iterate over all the keys.