Module Sapling.Nullifiers_hashed

type context = Raw_context.t * id

type elt = Tezos_protocol_environment_013_PtJakart.Sapling.Nullifier.t

The type of elements.

val mem : context -> elt -> (Raw_context.t * bool) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Tells whether an elt is a member of the set. Consumes Gas_repr.read_bytes_cost Z.zero

val init : context -> elt -> (Raw_context.t * int) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Adds an elt as a member of the set. Consumes Gas_repr.write_bytes_cost <size of the new value>. Returns the the new size.

val add : context -> elt -> (Raw_context.t * int * bool) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Adds an elt as a member of the set. Consumes Gas_repr.write_bytes_cost <size of the new value>. Returns the new size, and true if the value previously existed.

val remove : context -> elt -> (Raw_context.t * int * bool) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

Removes an elt from the set ; does nothing if not a member. Consumes Gas_repr.write_bytes_cost Z.zero. Returns the freed size, and a boolean indicating if a value was already associated to this key.

val fold_keys_unaccounted : context -> order:[ `Sorted | `Undefined ] -> init:'acc -> f:(elt -> 'acc -> 'acc Tezos_protocol_environment_013_PtJakart.Lwt.t) -> 'acc Tezos_protocol_environment_013_PtJakart.Lwt.t