Module Storage.Sc_rollup

module PVM_kind : Storage_sigs.Indexed_data_storage with type key = Sc_rollup_repr.t and type value = Sc_rollup_repr.Kind.t and type t := Raw_context.t

Smart contract rollup.

module Boot_sector : Storage_sigs.Indexed_data_storage with type key = Sc_rollup_repr.t and type value = string and type t := Raw_context.t

module Initial_level : Storage_sigs.Indexed_data_storage with type key = Sc_rollup_repr.t and type value = Raw_level_repr.t and type t := Raw_context.t

module Inbox : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.t and type value = Sc_rollup_inbox_repr.t and type t := Raw_context.t

module Last_cemented_commitment : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.t and type value = Sc_rollup_repr.Commitment_hash.t and type t := Raw_context.t

module Stakers : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Tezos_protocol_environment_013_PtJakart.Signature.Public_key_hash.t and type value = Sc_rollup_repr.Commitment_hash.t and type t = Raw_context.t * Sc_rollup_repr.t

module Staker_count : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.t and type value = int32 and type t := Raw_context.t

Cache: This should always be the number of entries in Stakers.

module Commitments : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.Commitment_hash.t and type value = Sc_rollup_repr.Commitment.t and type t = Raw_context.t * Sc_rollup_repr.t

module Commitment_stake_count : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.Commitment_hash.t and type value = int32 and type t = Raw_context.t * Sc_rollup_repr.t

Cache: This should always be the number of stakers that are directly or indirectly staked on this commitment.

module Commitment_added : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Sc_rollup_repr.Commitment_hash.t and type value = Raw_level_repr.t and type t = Raw_context.t * Sc_rollup_repr.t