Module Storage.Ticket_balance

This module exposes a balance table for tracking ticket ownership. The table is a mapping from keys to values where the keys consist of a hashed representation of:

A ticketer, i.e. the creator of the ticket

The content of a the ticket

The contract that owns some amount of the ticket The values of the table are the amounts owned by each key.

module Table : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type t := Raw_context.t and type key = Ticket_hash_repr.t and type value = Tezos_protocol_environment_013_PtJakart.Z.t

module Paid_storage_space : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Tezos_protocol_environment_013_PtJakart.Z.t

module Used_storage_space : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Tezos_protocol_environment_013_PtJakart.Z.t