Module Proof.Stream

Stream proofs represent an explicit traversal of a Merle tree proof. Every element (a node, a value, or a shallow pointer) met is first "compressed" by shallowing its children and then recorded in the proof.

As stream proofs directly encode the recursive construction of the Merkle root hash is slightly simpler to implement: verifier simply need to hash the compressed elements lazily, without any memory or choice.

Moreover, the minimality of stream proofs is trivial to check. Once the computation has consumed the compressed elements required, it is sufficient to check that no more compressed elements remain in the proof.

However, as the compressed elements contain all the hashes of their shallow children, the size of stream proofs is larger (at least double in size in practice) than tree proofs, which only contains the hash for intermediate shallow pointers.

type elt =

| Value of value

| Node of (step * kinded_hash) list

| Inode of hash inode

| Inode_extender of hash inode_extender

The type for elements of stream proofs.

Value v is a proof that the next element read in the store is the value v.

Node n is a proof that the next element read in the store is the node n.

Inode i is a proof that the next element read in the store is the inode i.

Inode_extender e is a proof that the next element read in the store is the node extender e.

type t = elt Tezos_protocol_environment_013_PtJakart.Seq.t

The type for stream proofs.

The sequance e_1 ... e_n proves that the e_1, ..., e_n are read in the store in sequence.