Module Tezos_raw_protocol_013_PtJakart.Ticket_balance_key

This module exposes a function for generating a ticket-balance key-hash given an owner and a ticket-token. The key-hash is used for populating the global ticket-balance table that tracks ownership of tickets for different tokens.

val of_ex_token : Alpha_context.context -> owner:Alpha_context.Destination.t -> Ticket_token.ex_token -> (Alpha_context.Ticket_hash.t * Alpha_context.context) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

of_ex_token ctxt ~owner ex_token returns the key_hash of the given owner and ex_token.