Module Tx_rollup_inbox_repr.Merkle

See Merkle_List for the documentation of those functions.

type tree

type root

type path

val empty : tree

val root : tree -> root

val (=) : root -> root -> bool

val compare : root -> root -> int

val root_encoding : root Tezos_protocol_environment_013_PtJakart.Data_encoding.t

val root_of_b58check_opt : string -> root option

val pp_root : Tezos_protocol_environment_013_PtJakart.Format.formatter -> root -> unit

val path_encoding : path Tezos_protocol_environment_013_PtJakart.Data_encoding.t

val add_message : tree -> Tx_rollup_message_hash_repr.t -> tree

val compute_path : Tx_rollup_message_hash_repr.t list -> int -> path Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult

val check_path : path -> int -> Tx_rollup_message_hash_repr.t -> root -> bool Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult

val path_depth : path -> int

val merklize_list : Tx_rollup_message_hash_repr.t list -> root

merklize_list messages construct a merkle root by build a tree, appending the messages one by one in the same order of the list and finally computing the root.