Module Tezos_raw_protocol_013_PtJakart.Tx_rollup_l2_context_sig

type signature = Tezos_protocol_environment_013_PtJakart.Bls_signature.signature

val signature_encoding : Tezos_protocol_environment_013_PtJakart.Bls_signature.signature Tezos_protocol_environment_013_PtJakart.Data_encoding.encoding

module Ticket_indexable : sig ... end

type address_index = Tx_rollup_l2_address.Indexable.index

An integer used to identified a layer-2 address. See Tx_rollup_l2_address.index.

type ticket_index = Ticket_indexable.index

An integer used to identified a layer-1 ticket deposited in a transaction rollup.

type metadata = {

counter : int64;

public_key : Tezos_protocol_environment_013_PtJakart.Bls_signature.pk;

}

The metadata associated to a layer-2 address.

The counter is an counter-measure against replay attack. Each operation is signed with an integer (its counter). The counter is incremented when the operation is applied. This prevents the operation to be applied once again, since its integer will not be in sync with the counter of the account. The choice of int64 for the type of the counter theoretically prevents the rollup to an integer overflow. However, it can only happen if a single account makes more than 1.8446744e+19 operations. If an account sends 1000 operations per seconds, it would take them more than 5845420 centuries to achieve that.

The public_key allows to authenticate the owner of the address, by verifying BLS signatures.

type Tezos_protocol_environment_013_PtJakart.Error_monad.error +=

| Balance_too_low

| Balance_overflow

| Invalid_quantity

| Unknown_address_index of address_index

| Metadata_already_initialized of address_index

| Too_many_l2_addresses

| Too_many_l2_tickets

| Counter_overflow

module type CONTEXT = sig ... end

This module type describes the API of the Tx_rollup context, which is used to implement the semantics of the L2 operations.