Module Tezos_raw_protocol_013_PtJakart.Tx_rollup_repr

This module defines identifiers for transaction only rollup (or tx rollup). It also specifies how to compute originated tx rollup's hash from origination nonce.

module Hash : sig ... end

A specialized Blake2B implementation for hashing tx_rollup identifiers with "txr1" as a base58 prefix

type t = private Hash.t

type tx_rollup = t

include Tezos_protocol_environment_013_PtJakart.Compare.S with type t := t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val in_memory_size : t -> Cache_memory_helpers.sint

in_memory_size tx_rollup returns the number of bytes tx_rollup uses in RAM.

val to_b58check : t -> string

val of_b58check : string -> t Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult

val of_b58check_opt : string -> t option

val pp : Tezos_protocol_environment_013_PtJakart.Format.formatter -> t -> unit

val encoding : t Tezos_protocol_environment_013_PtJakart.Data_encoding.t

val originated_tx_rollup : Origination_nonce.t -> t

originated_tx_rollup nonce is the tx_rollup address originated from nonce. See Origination_nonce.t for more information.

val rpc_arg : t Tezos_protocol_environment_013_PtJakart.RPC_arg.arg

module Index : Storage_description.INDEX with type t = t

val deposit_entrypoint : Entrypoint_repr.t

The entrypoint a layer-1 contract can use to deposit Michelson tickets into a transaction rollup.

module Set : Tezos_protocol_environment_013_PtJakart.Set.S with type elt = tx_rollup

module Map : Tezos_protocol_environment_013_PtJakart.Map.S with type key = tx_rollup