Module Tezos_raw_protocol_013_PtJakart.Tx_rollup_reveal_storage

val record : Raw_context.t -> Tx_rollup_repr.t -> Tx_rollup_level_repr.t -> message_position:int -> Raw_context.t Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

record ctxt tx_rollup lvl message_position adds message_position to the list of message with revealed withdrawals for tx_rollup at lvl.

val mem : Raw_context.t -> Tx_rollup_repr.t -> Tx_rollup_level_repr.t -> message_position:int -> (Raw_context.t * bool) Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

mem ctxt tx_rollup lvl message_position checks if message_position has already had its withdrawals revealed for tx_rollup at lvl.

val remove : Raw_context.t -> Tx_rollup_repr.t -> Tx_rollup_level_repr.t -> Raw_context.t Tezos_protocol_environment_013_PtJakart.Error_monad.tzresult Tezos_protocol_environment_013_PtJakart.Lwt.t

remove ctxt tx_rollup lvl clean-up the reveal accounting data from the layer-1 storage.